fix: crash in bc.py
